video output: add owner mechanism
Much like the audio_output and vlc_input_decoder, add a new owner
mechanism to provide interaction with the upstream resources using the
video output.

On audio output, start() provides a different aout_stream object, but we
always have the same object on video output. We could use vout_Start()
to install the owner but the window would already be enabled in that
case, so we use vout_Request() instead since this directly links to when
the input_decoder asks from the vout.

To be noted, there's still a weird situation with recycling as the video
output and window modules are still opened but the owner should not be
used since no client can actually make use of the reports. Since the
window is disabled at that moment, we actually ignore them. The locks
protecting this is also used to serialize the reports to the client.