configure.ac: add partial linking option
The option is disabled for now but will some day be enabled by default
since this is the only way to "support" static linking properly. It is
introduced here so that the option is used by following patches, but it
will really be useful only once we add supported platforms by the end of
the branch.