bin: vlc: add support for static partially linked plugin
This is the final test for the partial linking situation since we really
end up with a final binary that should have the static module bank
initialized correctly.

In the commit, we add the vlc_static_modules.c source file that contains
the standard definition for the static module bank and can be reused
against any binary being built to bundle the static module bank.