opengl: display: refactor with new render API
Rewrite the vout_display_t moduel to rely on a rendering function
matching the requirements of the new OpenGL provider API, and store the
rendering parameter into a dedicated structure to be compatible with
asynchronous rendering from the OpenGL provider in next commits.

Some care is needed regarding the picture and subpicture usage.

Co-author-by: Alaric Senat <alaric@videolabs.io>