opengl: add vlc_gl_RequestInit
The vlc_gl_RequestInit and vlc_gl_ReportInit APIs will allow the client
client to report that it is ready to initialize its renderer content, and
allow the OpenGL provider to select the best time and execution context
to initialize it.

The API will only be used "in sync mode" without dedicated thread for
the next commits, until every client is ported to the new API. Then,
vlc_gl_ReportInit() will be added and called from a dedicated thread.
Finally, asynchronous implementations will be able to decide the
execution flow by themselves while being notified that the client is
ready to change their content.