TODO opengl: add new async API
The vlc_gl_RequestRender and vlc_gl_ReportRender APIs allow the client
to report that its content can be refreshed with new data, and allow the
OpenGL provider to select the best time and execution context to start
the rendering of the new frame.

Size parameters are given since the size itself depends on the backing
storage which is managed by the OpenGL provider.