opengl: add async request changes function
When a change in renderer is requested, like UpdateFormat from the
OpenGL display, this provides a way for the setup callback to be
executed on the correct thread.

Currently, this executes the callback in-place since there is no async
implementation nor dedicated thread, but this will allow switching to a
dedicated thread in next commits.