opengl: add owner to offscreen GL
The owner represents the client of the OpenGL provider instance, which
will configure how the renderer is initialized and how the rendering is
done for the surface.

Technically, most OpenGL provider implementations for filters are
correctly working with synchronous implementations but emscripten
implementation will also fail to share the OpenGL state across different
threads.