win32: window: avoid closing window on WM_CLOSE
We don't want to close the window on WM_CLOSE, since it would break the
vout display relying on the graphical context of the window to do the
rendering.

Instead, ReportClose should disable/close the window after the display
is closed correctly, since we now have the proper propagation of the
event.

Prevent crash with wgl when closing.