vlcrs: add meson infra for building and linking Rust modules
Despite meson officially "supporting" Rust it does support working with
Cargo (Rust package manager and build system) so we need to have
a bit of custom system for using it. `cargo-output.py` to build and copy
the output files (static archive and depfile) to the expected location
by meson. And `cargo-rustc-static-libs.py` to get the common linker args
for a bar-bone std rust program (ie without dependencies).

Those two scripts would potentially be removed if meson and/or cargo
became more flexible and/or intelligent.