input: decoder: start clock when decoders are ready
DecoderWaitUnblock is waiting for the vlc_input_decoder_StopWait which
will be called by the es_out when every decoders are ready to start
playing.

When DecoderWaitUnblock is called and almost ready to return, it means
that we already have a frame in the output and that playback has
started. We can compute the monotonic start by calling vlc_clock_Start.

Some care must be taken given that other reasons might unblock the
decoders, like flushing or closing the decoder.

Currently, vlc_clock_Start() does nothing as long as we didn't start the
input vlc_clock_t also, but all of the code will be enabled after.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>