TODO: es_out: remove buffering_duration from start_time
TODO: Fix the choice of PCR to after preroll and add tests for preroll
      and extra buffer situations. The change on input_dejitter is
      currently breaking the test checking that input_dejitter is
      correctly accounted.

When the es_out was first written, the start_time was directly set as
the current time minus the pts_delay when starting the playback,
highlighted in the following commit:

    commit e5ed591a2e074ca07dd7d34f207834a02552a2f6
    Author: Laurent Aimar <fenrir@videolan.org>
    Date:   Wed Oct 8 21:20:05 2008 +0200

        Implemented buffering manipulation at the es_out level.
        
        It allows faster seek and start up time when the access pace can be controlled.

It introduce the EsOutDecodersStopBuffering() function and the other
call to the function input_clock_ChangeSystemOrigin() as well as the
decoder wait mechanism.

    +    input_clock_ChangeSystemOrigin( p_sys->p_pgrm->p_clock,
                        i_ts_delay - p_sys->p_input->i_pts_delay );

Preroll was added in the following commit, so as to implement the first
frame display as soon as possible.

    commit 4d4f76fd38af44ee1bd99bb221c8201b4569c1f9
    Refs: 0.9.0-1338-g4d4f76fd38a
    Author:     Laurent Aimar <fenrir@videolan.org>
    AuthorDate: Fri Oct 10 20:39:24 2008 +0200
    Commit:     Laurent Aimar <fenrir@videolan.org>
    CommitDate: Fri Oct 10 21:28:30 2008 +0200

        Partially fixed preroll support.

Then the next commit added buffering_duration to the clock stream to
system computation, to handle extra buffering:

    commit 2f9d9766dfe25f100ef04d0d812aebdd18401c0d
    Refs: 1.0.0-rc1-1295-g2f9d9766dfe
    Author:     Laurent Aimar <fenrir@videolan.org>
    AuthorDate: Thu Jul 16 14:50:16 2009 +0200
    Commit:     Laurent Aimar <fenrir@videolan.org>
    CommitDate: Thu Jul 16 21:12:44 2009 +0200

        Let the input bufferize more data when possible.

         The input will try to read 20% faster the source until a certain amount
        of data is buffered by the fifo of the decoders (for now 10Mbytes or
        500kbytes when OPTIMIZE_MEMORY is defined for the sum of all fifos).
         This buffering adds up to pts_delay without any additional delay but can
        only work when VLC controls the source pace.

         It has a drawback with the current way the meta data works: they are seen
        too soon.

Then the next frame support commit adds the buffering_extra variables:

    commit b0782ad7dcfd29f942098e3160ced83a9ade6dc7
    Refs: 0.9.0-1565-gb0782ad7dcf
    Author:     Laurent Aimar <fenrir@videolan.org>
    AuthorDate: Wed Oct 22 21:55:06 2008 +0200
    Commit:     Laurent Aimar <fenrir@videolan.org>
    CommitDate: Wed Oct 22 22:21:44 2008 +0200

        Added "next frame" support.

        It will work most of the time.
        Subtitle are not correctly handled.
        It will be ignored if no video.

Finally, the following commit added the track_pts_delay

    commit 1f51e1d214771f7d609d518178c0e4c173dff8dd
    Author: Thomas Guillem <thomas@gllm.fr>
    Date:   Tue Jun 18 14:14:45 2019 +0200

        input: rework delay handling
    
All of those additions were built on top of the input_clock changes,
which is in particular shifting the playback timestamps pts_delay in
the future, and on top of the preroll-like changes (additional buffering
at the beginning) and extra-buffering-like changes (additional buffering
at the end) to calibrate the beginning of the stream on the first PCR.
For instance, if the first PCR is part of the preroll, then it should be
in the past so that the current time is actually the preroll end.

          first     End of     Input
           PCR      Preroll   Start PCR
            v           v        v
    |-------x-----------x--------x---> media time
            |--------------------|     Buffering duration
                        |--------|     pts_delay

            <--------------------|     System time is shifted back
                                       as if it was played with
                                       monotonic clock rate=1

Instead in this commit, we try to setup the start of the stream to the
correct date directly.

          first     End of     Input
           PCR      Preroll   Start PCR
            v           v        v
    |-------x-----------x--------x---> media time
            |--------------------|     Buffering duration
                        |--------|     pts_delay

            |---------->               Media time will be shifted forward
                                       to skip the preroll part, there's
                                       no need to handle extra-buffering.

This will ease the understanding of the clock start and simplify the
interaction with clock drift when playing live media, and rate change
when rate != 1.