service_discovery: upnp: fix interface selection on linux and android
Previously, the interface selection was done by libupnp, which just
selected the first it found, which may or may not connected.

The new implementation actually selects the first one that gets an ssdp
answer. We don't use netlink api because Android has strong SElinux
policy preventing their use. This is probably the most portable way to
do this.