configure: luac: use host variable from with-contrib processing
Currently, autodetection of luac in contrib fails, because:

When building the luac contrib, it is built at
$(PREFIX)/../bin/$(cc -dumpmachine)-luac

When configuring, if --host is not provided, the script looks for luac
at $(PREFIX)/../bin/-luac as $host_alias is empty, failing to lookup
luac.

Use (cc -dumpmachine) to look for the LUA byte compiler when host is not
provided.

This is already processed at with-contrib proccessing section. Make that
global so that 7d31c85c's improvement do not break and use that.