vlcrs-tick: make conversion and operations faillible
Overflows can reasonably happen with ticks so callers must be warned and
actually handle the overflow themselves. Those overflows can happen when
converting from another unit, because of the CLOCK_FREQ quantization, or
when arithmetics operations are used.

A MaybeOverflow<T> context is setup when arithmetics operations are
used, which allows operations to be done together before the result is
checked at the end.