vlcrs: tick: remove bindgen usage for vlcrs-tick
No consensus was reached for using bindgen on vlcrs crates so we'll
remove the bindgen usage for now from the vlcrs-sys-generator.

It means that we cannot maintain the documentation tags, which risk
being out-of-date so we'll also remove them. A future issue will track
and explore how to forward documentation from the C API.