clock: implement vlc_clock_Start
The es_out is handling the current startup sequence for the clock by
resetting the clock, shifting the input clock and setting the startup
time again. Those functions are mixing three different components and
behaviour and makes it hard to test the startup scenarios.

vlc_clock_Start() on the input vlc_clock_t will replace those calls by
providing the correct offset on the output points instead of
transforming the input points.

In addition, vlc_clock_Start() will be provided to the output clocks so
that they can provide the wait_sync reference point, instead of relying
on the first conversion done.

Like first_pcr, start_time must be shifted when pause is set. It would
also be possible to "reset" the start_time when un-pausing to the
un-pause date.

Co-authored-by: Thomas Guillem <thomas@gllm.fr>