vlcrs-core: extension: Refactor Resource Management and Add Cloning
Removed the use of ManuallyDrop and introduced a release function for controlled
deallocation of resources in the Extension struct. Added Clone implementation to
enable creating new Extension instances from the same raw pointer.

Note: Deallocation of Extension resources is managed exclusively by ExtensionManager,
ensuring it is safe to clone Extension instances with the raw pointer.