wasm-rs: configurations: add function to write list of paths to wasm memory
Previously, the mechanism for handling paths involved joining the path
strings with a ; separator. This update introduces support for writing
a list of strings directly to WebAssembly memory. The list is composed
of pointers to the strings stored in memory, with the end of the list
indicated by a null pointer.