d3d11_fmt: allow using D3D11 by default on all systems
On vanilla Win7 or with old drivers the feature level provided may be lower than 11.1.
We can now deal with it in the display module. We convert NV12/P010 to RGB
in the GPU.

The shader model levels are already handled. The display module was working fine on
Win7 but the decoder format was causing issues.