qml: use `text/uri-list` for internal dragging instead of input items
We already handle `text/uri-list`, since we support dragging from
external sources. However, internally, this route is not used and
rather input items are considered for dragging.

This change does not affect internal intra-view move, but affects
inter-view insertion.

This also makes it possible to drag media library items that are
collection of media outside the application.