qt: do not take pixel ratio into consideration when adjusting application scale factor
The application scale factor is used in `dp()`, which is in most cases if not all
used as item size in QML.

Pixels in QML are not device independent, such that Qt applies the device pixel ratio
to them. This has some exceptions, notably `sourceSize` of `Image`.

Applying device pixel ratio twice would not yield the intended results.