opengl: add vlc_gl_RequestRender API / commit message
The vlc_gl_RequestRender and vlc_gl_ReportRender APIs will allow the
client to report that its content can be refreshed with new data, and
allow the OpenGL provider to select the best time and execution context
to start the rendering of the new frame.

Size parameters are given since the size itself depends on the backing
storage which is managed by the OpenGL provider.

The API will only be used "in sync mode" without dedicated thread for
the next commits, until every client is ported to the new API. Then,
vlc_gl_ReportRender() will be added and called from a dedicated thread.
Finally, asynchronous implementations will be able to decide the
execution flow by themselves while being notified that the client is
ready to change their content.