opengl: display: use async changes API for UpdateFormat
The vlc_gl_RequestChanges API is async-mode compatible, and requires the
use of a dedicated data structure and callback to potentially move the
execution context to another thread depending on the implementation.

The modules need this callback to synchronously return whether it could
apply the changes or not, but the OpenGL module is not re-entered so
other operations can be used from the callback itself if needed.