qt: assume touchpad wheel events are inverted if invertedness is not reported in `WheelToVLCConverter`
`WheelToVLCConverter` is not used for scrolling, so we have to handle inversion.
This was already done in 1ba2a2db, but unfortunately not all platform (plugins)
report invertedness.

According to my knowledge, Mac and Wayland platform plugins report, while Xcb
and Windows/Direct2D do not report.

Usually touchpad scrolling direction is inverted, which may be the source of
events in this class, so it makes more sense to assume the wheel events are
inverted by default.