qt: model_recovery_agent: transform exception into result
There is no need for an unwinding/asynchronous mechanism for the error
handling since the result was caught directly at the caller site in a
synchronous way. As a matter of fact, the catch block was even empty.

For an unknown reason, throwing from the ModelRecoveryAgent class was
also triggering the unwinding mechanism despite the try/catch block.

This uses a static initializer function instead of using the constructor
directly, so as to handle the error beforehand.

Fixes #29071