WIP: mediacodec: remove timestamp_Fifo
timestamp_Fifo is used to work around the lack of PTS on our side.
MediaCodec documentation says that it need PTS and no DTS.

We should find a way to always have valid PTS on the decoder side, by
triggering a packetizer or fixing demuxers that says the fmt is already
packetized.

timestamp_Fifo should be removed as it can't handle silent frame drops
from MediaCodec, that could lead to issues with the next commit.