mediacodec: try to calculate missing PTS
From the MediaCodec documentation
The presentation timestamp in microseconds for this buffer. This is
normally the media time at which this buffer should be presented
(rendered). When using an output surface, this will be propagated as the
timestamp for the frame (after conversion to nanoseconds).

So we should never pass DTS here as it can decrease and increase. That's
was what the timestamp_Fifo was working arround. So instead, in case of
missing PTS, calculate the PTS from the first DTS.