qt: do not construct `ModelRecoveryAgent` if settings or playlist controller do not exist
`ModelRecoveryAgent` requires valid settings and playlist controller instances. It can
not be constructed when they do not exist. This is an extremely unlikely case (but not
impossible), because given the structure of the application it is expected that they
are valid when `ModelRecoveryAgent` is constructed.

It should be noted that the reason `ModelRecoveryAgent` construction is delayed through
queued connection is only to not prolong the application startup time.