contrib/src: use HTTPS for most contribs
This patch updates the URLs for contrib retrieval from
HTTP to HTTPS, where possible. URls that could not be
moved to HTTPS are marked with an XXX comment.

One domain is additionally changed: libsamplerate is
now retrieved from the libsndfile organization
on GitHub instead of mega-nerd.com, as the latter
does not offer HTTPS. The former was confirmed
for release integrity against the latter and was
cross-checked as the living source for libsamplerate
against Homebrew and Arch Linux, both of which use
theGitHub organization.

Signed-off-by: William Woodruff <william@trailofbits.com>
(cherry picked from commit f143a3b131107a950b812d276e778a3e17bae96b) (edited)
edited:
- some contribs don't exist in VLC 3 (basu, librist)
- live555 is using the VLC server exclusively (9af12a6e8240b89da4d9c2ff7e031bff9fc6c829)
- GNU is already using ftpmirror.gnu.org
- some contrib versions have changed since then
Signed-off-by: Steve Lhomme <robux4@ycbcr.xyz>