qml: Update MusicArtist view
The view header is moved out of the table/ǥrid internal header, the main goal is
to allows detaching the control bar when scrolling. This also makes the
navigation between the list/grid and the prelude simpler as the items are now
siblings and not parent/child

Now the view uses a system of nested flickable, the grid/list flickable is embed
in a flickable that also contains headers and other components, the position of
the list/grid flickable is updated according to the master
flickable. Autoscrolling from the inner flickable is propagated to the master
flickable