bluray: use an atomic refcounter to release the bluray_overlay_t
The region updater sys is now the bluray_overlay_t with its placed pictures.
There is one bluray_overlay_t per plane (2 planes). It may be still
referenced by the vout when the bluray is not using anymore. Subsequent
calls to subpictureUpdaterUpdate() should not render anything and
subpictureUpdaterDestroy() will finally release it.