qml: adjust opacity in delegate in `ControlRepeater`
Unlike visibility, this does not need to be delayed.
This is useful to not display the control if in the
first pass it should not be displayed.

For example, when the window has small width and the
user switches pages, the controls do not appear just
to disappear in the second pass, thanks to this.