qml: do not fully collapse the artists pane in `MusicArtistsAlbums`
When there is not enough width, we can collapse to just show the
cover art. This is also the planned behavior of the upcoming side-
bar (showing an icon when available, or abbreviation, instead of
fully collapsing).

Only showing the cover art requires less than 50dp width. Requiring
such small space makes full collapsing hard to justify.

Fully collapsing things are incredibly bad for UX, especially when
there is no alternative to what the collapsed thing provides. In
this case, the user has no option changing the current item when
the artists pane is fully collapsed.