qt: introduce `LimiterProxyModel`
This is a simple class that derives from `QSortFilterProxyModel`.
It provides properties `maximumRowCount` and `maximumColumnCount`
to make it easier to filter an item model based on arbitrary
thresholds.

The advantage of this is that `QSortFilterProxyModel` is clever
enough to optimize emitting signals. It does not emit model
reset signal in each change (to make it job much easier). Instead,
it signals only the part of the model that changes. For example,
if the row limit increases from 4 to 5, provided that the source
model accompanies that, it signals `QAbstractItemModel::rowsInserted()`
with the appropriate indexes.

This makes it very useful (as its name suggests) for filtering
and sorting purposes because this way the view does not need to
reload all items.