qml: add property `viewModel` to `MediaView`
By default it is bound to the main model.

This is necessary because in `MainViewLoader`,
model type is `BaseModel`, which bars models
that derive from `QAbstractItemModel` but not
from `BaseModel`, such as proxy models, to be
usable there.

I don't get why `BaseModel` derives from
`QAbstractItemModel`. It could be a model
interface that final models derive from both
that interface and also an abstract model so
that we could use `QAbstractItemModel` as the
base type instead, and not have this problem.