qml: use `LimiterProxyModel` in `HomePage`
This fixes, to considerable extent, extremely
frustrating resize behavior in the home page.
Particularly observed when adjusting the width
of the play queue pane, or the window width.

Currently the view models emit model reset signal
as soon as the limit changes. This causes the
views to re-load all items, which is not an
acceptable behavior.

I wanted to use a screen-related constant for
the maximum limits (similar to what we are doing
in the player view to determine the cover art's
maximum size), but I have not found such
information provided by `ExpandGridItemView`.