aout: srt: use a define instead of a const int
Use a define of the same value (and static assert that it matches)
instead of a define.

Even though it is const int, that doesn't make it an integer constant
so strictly following the specification, the array would be a variable
length array.

Instead use a define which is the best thing we have until we get a
constexpr in C23, solving the following warning:

  variable length array folded to constant array as an extension