qt: use type-safe connections in most places
Most signal-slot connects are rewritten from the old string-based runtime
selection to the compile-time selection introduced in Qt 5. This is in spirit
effectively a backport of [1], but written from scratch due to excessive
conflicts due to code divergence. This will detect some of the issues mentioned
in [2] (at least the "Open Recent Media" doing nothing), due to detecting the
removal of the slots at compile time (and thus failing to compile, instead of
producing a binary that doesn't work).

This addresses the immediate usages of CONNECT or SIGNAL/SLOT macros only.
Usages through other macros will be addressed separately for clarity.

[1] https://code.videolan.org/videolan/vlc/-/commit/eb0b785fdf5396b429fb039b0c92d44f768fe65c
[2] https://code.videolan.org/videolan/vlc/-/merge_requests/7754#note_499875