contrib: libarchive: disable auto dependencies on Darwin
When building on Darwin, CMake checks for dependencies in non-standard
locations like the Homebrew and MacPorts installations without any easy
way to opt out of this, apparently. In addition, it does not properly
set the linker flags to these paths if it end up finding a library in
there so it just results in a linker error in the end.

To work around this, disable all autodetected dependencies that we do
not provide as contribs.