codec: videotoolbox: fix cropping for yuv420p on OSX
CMVideoFormatDescriptionCreate was being passed i_visible_width/height
instead of i_width/i_height. Meanwhile the output buffer was being
created with the full width/height. This told VideoToolbox the wrong
bitstream dimensions when the two aren't equal (e.g. MKV container
crop).

This results in decoding a subset of the video, then scaling it up to
the full size. Then when the image is displayed it gets cropped/scaled
again as a texture, with very incorrect results. This fixes it by
passing the actual encoded frame dimensions.